Illustration for this file:
This is a program that aims to test whether there's a structural change of a stock's return (%) in the beginning of the month and in the end of the month. The professor teached this in Excel, and I made a python version to make the process easier. Basically you can download data from Yahoo Finance and get whatever stock you want to test that.